refactor(trace): extract QueueEventList, rewrite picker with payload panel
- New standalone QueueEventList component (features/queue-event-list/)
  Replaces all fetch/search/validation/keyboard-nav logic from
  queue-events-tab. Accepts searchAnchorTime, getSchemaFn, latestEid,
  latestWriteMs, onTrace, onReplay, onEventSelect, shareUrlFn props so
  it works in any context without AppState dependency.

- queue-events-tab.svelte slimmed to a thin dashboard wrapper (~20 lines)
  Passes timePicker.endTime as searchAnchorTime, compState.getSchema as
  getSchemaFn, and settings for jump-to-latest.

- trace-queue-picker.svelte rewritten using QueueEventList + SplitPane
  Row click selects an event and opens a right-panel payload viewer.
  Zap button navigates to the trace. Payload panel includes CodeView /
  DiffCodeView (old/new toggle), copy button, and S3 link detection.

Task: ES-3029
