+++ webapp/src/lib/client/components/features/queue-event-list/index.ts
@@ -1,0 +1,2 @@
+export { default as QueueEventList } from './queue-event-list.svelte';
+export type { StreamEvent } from './queue-event-list.svelte';
